Use size_t instead of int for ptrace_data_t

+++ libr/include/r_io.h
@@ -51,7 +51,7 @@ typedef void * r_ptrace_data_t;
 #define R_PTRACE_NODATA NULL
 #else
 typedef int r_ptrace_request_t;
-typedef int r_ptrace_data_t;
+typedef size_t r_ptrace_data_t;
 #define R_PTRACE_NODATA 0
 #endif
 #endif
